pyproject.toml
[tool.poetry]
name = "grocery-list-service"
version = "0.1.0"
description = "FastAPI backend for Grocery List management"
authors = ["MobyNL <markmoberts@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pyjwt = "^2.10.1"
pydantic = "^2.10.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
